SignUp form

+++ src/main/java/Smallcare/Controllers/PetController.java
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -20,10 +19,8 @@ import javax.persistence.EntityNotFoundException;
 
 import java.io.File;
 import java.io.*;
-import java.util.List;
 
 
-@RequestMapping("/user/pets")
 @Controller
 public class PetController {
 
@@ -36,26 +33,30 @@ public class PetController {
     @Value("${upload.path}")
     private String upload_path;
 
-    @GetMapping("")
-    public String pets(Model model) {
+    private User getCurrentUser(){
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         if (auth instanceof AnonymousAuthenticationToken) {
-            return "index";
+            return null;
+        }
+        return ((User) auth.getPrincipal());
+    }
+    @GetMapping("/pets")
+    public String pets(Model model) {
+        User user = getCurrentUser();
+        if(user == null){
+            return "/index";
         }
         Iterable<Pet> petList = userService
                 .findById(
-                        ((User) auth
-                                .getPrincipal())
-                                .getId())
+                        (user).getId())
                 .getPetList();
-        System.out.println(petList != null);
         if (petList != null) {
             model.addAttribute("pets", petList);
         }
         return "pets";
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("/pets/{id}")
     public String pet(Model model, @PathVariable Long id) {
         Pet pet = petService.findById(id).orElseThrow(EntityNotFoundException::new);
         if (pet != null) {
@@ -64,49 +65,49 @@ public class PetController {
         return "pet";
     }
 
-    @GetMapping("/add")
+    @GetMapping("/pets/add")
     public String addPet(Model model) {
         model.addAttribute("pet", new Pet());
         return "addPet";
     }
 
-    @PostMapping("/add")
+    @PostMapping("/pets/add")
     public String postPet(Model model,
                           @ModelAttribute Pet pet,
-                          @RequestParam(value = "file", required = false) MultipartFile file) throws IOException {
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if (auth instanceof AnonymousAuthenticationToken) {
-            return "index";
+                          @RequestParam(value = "file", required = false) MultipartFile file) {
+        User user = getCurrentUser();
+        if(user == null){
+            return "/index";
         }
         if (pet.getName() != null) {
-            User user = (User) auth.getPrincipal();
             Pet newpet = petService.save(pet);
             userService.addPet(user, pet);
             if (file != null) {
-                file.transferTo(new File(upload_path + newpet.getId().toString() + ".png"));
+                try {
+                    file.transferTo(new File(upload_path + newpet.getId().toString() + ".png"));
+                } catch (IOException ioException){
+                    System.out.println("Bad file saving :(");
+                }
             }
         }
-
         Iterable<Pet> petList = userService
                 .findById(
-                        ((User) auth
-                                .getPrincipal())
-                                .getId())
+                        (user).getId())
                 .getPetList();
-        System.out.println(petList != null);
         if (petList != null) {
             model.addAttribute("pets", petList);
         }
         return "pets";
     }
 
-    @GetMapping("/delete/{id}")
-    public String deletePet(Model model, @PathVariable Long id) {
-        petService.deleteById(id);
-        List<Pet> pets = petService.findAll();
-        if (!pets.isEmpty()) {
-            model.addAttribute("pets", pets);
+    @GetMapping("/pets/delete/{id}")
+    public String deletePet(Model model, @PathVariable Long id) throws Exception {
+        User user = getCurrentUser();
+        if(user == null){
+            return "/index";
         }
-        return "redirect:/user/pets";
+        userService.deletePet(user, petService.findById(id).orElseThrow(Exception::new));
+        petService.deleteById(id);
+        return pets(model);
     }
 }
